route_optimizer/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.4
gunicorn>=23.0
requests>=2.31
polyline>=2.0        # Decode Mapbox-encoded polyline5 geometry
scipy>=1.12          # KDTree for O(log n) spatial queries
numpy>=1.26
python-dotenv>=1.0
